requirements.txt
api==0.0.7
asgiref==3.2.3
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
more-itertools==8.2.0
nose==1.3.7
packaging==20.1
Pillow==7.0.0
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
pytest-django==3.8.0
pytz==2019.3
requests==2.23.0
selenium==3.141.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
wcwidth==0.1.8
zipp==3.0.0
